# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED, INVALID_TOKEN, CONFLICT]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        origin_type:
          type: string
          nullable: true
          description: '"imported" when the listing is imported (else null) — drives the "مستورد / Imported" card badge on industrial + car listings.'
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
          description: Additive. The listing's main section — lets clients adapt per-section UI (e.g. the save glyph) without re-deriving it.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
        is_bookable:
          type: boolean
          nullable: true
          description: >-
            Additive. True when this listing is a furnished/daily rental (hotel
            model) that can be reserved from its detail. Lets the feed, search
            card, and map pin surface a "قابل للحجز / Bookable" affordance without
            fetching the detail. Null/false for everything else.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    DealerListing:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        price_display:
          type: string
        price_raw:
          type: string
        location:
          type: string
        status:
          type: string
        created_at:
          type: string
        views:
          type: integer
        clicks:
          type: integer
        leads:
          type: integer
    Lead:
      type: object
      properties:
        id:
          type: string
        listing_id:
          type: string
        listing_title:
          type: string
        action_type:
          type: string
          enum: [whatsapp, call, chat, finance_request]
        status:
          type: string
          enum: [new, contacted, closed]
        buyer_name:
          type: string
          nullable: true
        buyer_phone:
          type: string
          nullable: true
        created_at:
          type: string
...
    DealerStats:
      type: object
      properties:
        active_listings:
          type: integer
        total_listings:
          type: integer
        leads_today:
          type: integer
        conversion_rate:
          type: string
        total_views:
          type: integer
        leads_chart:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              leads:
                type: integer
    BoostResult:
      type: object
      properties:
        ad_id:
          type: string
        listing_id:
          type: string
        ad_type:
          type: string
        expires_at:
          type: string
        duration_days:
          type: integer
        promo_used:
          type: string
          description: >
            Promo ad credit applied to this boost (consumed before the real
            wallet), as a 2-dp string. "0.00" when no promo was available.
        wallet_charged:
          type: string
          description: >
            Real wallet money charged — the remainder after promo credit — as a
            2-dp string.
    ImportResult:
      type: object
      properties:
        success_count:
          type: integer
        failed_count:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              batch_index:
                type: integer
              rows_in_batch:
                type: integer
              message:
                type: string
...
    CompanyDirectoryItem:
      type: object
      description: >
        Additive (Task #40). Suppliers directory entry with real active-listing
        and follower counts. is_following is viewer-relative (false when
        unauthenticated).
      required:
        - id
        - name
        - role
        - is_verified
        - industry
        - hq_country
        - logo_url
        - cover_url
        - active_listings
        - follower_count
        - is_following
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        is_verified:
          type: boolean
        industry:
          type: string
          nullable: true
        hq_country:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
        cover_url:
          type: string
          nullable: true
        active_listings:
          type: integer
        follower_count:
          type: integer
        is_following:
          type: boolean
    FollowResult:
      type: object
      description: Additive (Task #40). Result of a follow/unfollow toggle.
      required: [following, follower_count]
      properties:
        following:
          type: boolean
        follower_count:
          type: integer
...
paths:
...
  /v1/dealer/stats:
    get:
      operationId: getDealerStats
      summary: Dealer dashboard stats
      tags: [dealer]
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/DealerStats"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/dealer/listings:
    get:
      operationId: getDealerListings
      summary: Dealer's own listings
      tags: [dealer]
      security:
        - ClerkAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, sold, archived]
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at, price, views, leads]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DealerListing"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/dealer/listings/boost:
    post:
      operationId: boostListing
      summary: Boost listing as sponsored ad
      tags: [dealer]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id]
              properties:
                listing_id:
                  type: string
                ad_type:
                  type: string
                  enum: [featured, native_feed, top_search]
                duration_days:
                  type: integer
                  default: 7
                idempotency_key:
                  type: string
                  description: >
                    Optional client-supplied key. Retrying a boost with the same
                    key returns the original ad without re-consuming promo credit
                    or re-charging the wallet.
      responses:
        "200":
          description: Boost created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BoostResult"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/dealer/listings/import:
    post:
      operationId: bulkImportListings
      summary: Bulk import from CSV
      tags: [dealer]
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          text/csv:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImportResult"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/dealer/leads:
    get:
      operationId: getDealerLeads
      summary: Dealer CRM leads
      tags: [dealer]
      security:
        - ClerkAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, contacted, closed]
        - in: query
          name: action_type
          schema:
            type: string
            enum: [whatsapp, call, chat, finance_request]
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/dealer/leads/{id}:
    patch:
      operationId: updateLeadStatus
      summary: Update lead status
      tags: [dealer]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [new, contacted, closed]
      responses:
        "200":
          description: Updated

  # ── PUBLIC: user-submitted reports & support tickets ──────────────────
...
  /v1/companies/{id}/listings:
    get:
      operationId: getCompanyListings
      summary: Public active listings for a company (Task #33)
      tags: [companies]
      security:
        - {}
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated company listings (FeedItem contract)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...
  /v1/companies/{id}/follow:
    post:
      operationId: followCompany
      summary: Follow a company / supplier (Task #40)
      tags: [companies]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow state after toggle
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/FollowResult"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
    delete:
      operationId: unfollowCompany
      summary: Unfollow a company / supplier (Task #40)
      tags: [companies]
      security:
        - ClerkAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow state after toggle
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/FollowResult"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
  /v1/me/following:
    get:
      operationId: listMyFollowing
      summary: Companies the caller follows (Task #40)
      tags: [companies]
      security:
        - ClerkAuth: []
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated followed companies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CompanyDirectoryItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"